+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/Network.kt
@@ -1,0 +1,9 @@
+package com.atlassian.performance.tools.awsinfrastructure
+
+import com.amazonaws.services.ec2.model.Subnet
+import com.amazonaws.services.ec2.model.Vpc
+
+internal class Network(
+    val vpc: Vpc,
+    val subnet: Subnet
+)